feat(i18n): internationalization and localization improvements
This commit introduces the following enhancements to the internationalization and localization features of the application: - Chore: Changed the flag for the English language from UK to US. - Feat: Added support for the Portuguese language (pt-BR). - Chore: Created initial translation files for Portuguese. - Feat: Updated the language labels in the language selector to their native names.